+++ android/app/src/main/java/com/mulberry/ody/presentation/room/detail/listener/InviteCodeCopyListener.kt
@@ -1,0 +1,5 @@
+package com.mulberry.ody.presentation.room.detail.listener
+
+interface InviteCodeCopyListener {
+    fun onCopy()
+}
