+++ android/app/src/main/java/com/mulberry/ody/presentation/room/log/NotificationLogFragment.kt
@@ -37,18 +37,5 @@ class NotificationLogFragment :
         collectWhenStarted(viewModel.notificationLogs) {
             notificationLogsAdapter.submitList(it)
         }
-        collectWhenStarted(viewModel.copyInviteCodeEvent) {
-            viewModel.copyInviteCodeEvent.collect {
-                val intent = Intent(Intent.ACTION_SEND_MULTIPLE)
-                intent.type = "text/plain"
-
-                val shareMessage =
-                    getString(R.string.invite_code_copy, it.meetingName, it.inviteCode)
-                intent.putExtra(Intent.EXTRA_TEXT, shareMessage)
-
-                val chooserTitle = getString(R.string.invite_code_copy_title)
-                startActivity(Intent.createChooser(intent, chooserTitle))
-            }
-        }
     }
 }
