+++ android/app/src/main/res/layout/item_invite_code_copy.xml
@@ -4,6 +4,9 @@
 
     <data>
 
+        <variable
+            name="inviteCodeCopyListener"
+            type="com.mulberry.ody.presentation.room.detail.listener.InviteCodeCopyListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -41,7 +44,8 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/tv_invite_code_guide"
-            app:layout_constraintVertical_chainStyle="packed" />
+            app:layout_constraintVertical_chainStyle="packed"
+            app:onSingleClick="@{() -> inviteCodeCopyListener.onCopy()}" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
